refactor: move Modal component to separate folder
